requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
sentence-transformers==2.2.2
chromadb==0.4.15
openai==1.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
pdfplumber==0.10.3
PyPDF2==3.0.1
pytesseract
pdf2image
pillow